remove ground_map_1.c
