+++ ld_script.txt
@@ -341,7 +341,6 @@ SECTIONS {
         src/code_80A26CC.o(.text);
         asm/code_80A26CC_1.o(.text);
         src/ground_map.o(.text);
-        src/ground_map_1.o(.text);
         asm/ground_map_1.o(.text);
         src/ground_map_2.o(.text);
         src/ground_sprite.o(.text);
